[design] 개인액션리스트 ui수정
