+++ components/header/HeaderProfile.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const HeaderProfile = () => {
+  return <div>HeaderProfile</div>;
+};
+
+export default HeaderProfile;
